refactor: simpler original impl
